# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Siklas API Documentation",
        "description": "API documentation for the Siklas project",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/admin/dosen": {
            "get": {
                "tags": [
                    "Admin - Dosens"
                ],
                "summary": "List and search lecturers",
                "operationId": "listDosens",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List dosen berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Dosens"
                ],
                "summary": "Create a new lecturer",
                "operationId": "storeDosen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDosenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dosen berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/dosen/{id}": {
            "put": {
                "tags": [
                    "Admin - Dosens"
                ],
                "summary": "Update an existing lecturer",
                "operationId": "updateDosen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDosenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dosen berhasil diperbarui"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Dosens"
                ],
                "summary": "Delete a lecturer",
                "operationId": "deleteDosen",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dosen berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/user": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "List and search users",
                "description": "Retrieve a paginated list of users with search and role filters",
                "operationId": "listUsers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filter by role",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "mahasiswa",
                                "kaprodi",
                                "dosen"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Create a new user",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/user/{id}": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Update an existing user",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Delete a user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreDosenRequest": {
                "required": [
                    "nama",
                    "bidang_penelitian_major_id"
                ],
                "properties": {
                    "nama": {
                        "type": "string",
                        "example": "Dr. John Doe"
                    },
                    "gelar_awal": {
                        "type": "string",
                        "example": "Dr."
                    },
                    "gelar_akhir": {
                        "type": "string",
                        "example": "M.Kom."
                    },
                    "bidang_penelitian_major_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "minors_id": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
            "UpdateDosenRequest": {
                "properties": {
                    "nama": {
                        "type": "string",
                        "example": "Dr. John Doe Updated"
                    },
                    "gelar_awal": {
                        "type": "string",
                        "nullable": true
                    },
                    "gelar_akhir": {
                        "type": "string",
                        "nullable": true
                    },
                    "bidang_penelitian_major_id": {
                        "type": "integer"
                    },
                    "minors_id": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "required": [
                    "name",
                    "username",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "mahasiswa",
                            "admin",
                            "kaprodi",
                            "dosen"
                        ],
                        "example": "mahasiswa"
                    },
                    "program_studi": {
                        "type": "string",
                        "enum": [
                            "PTI",
                            "SI",
                            "TI",
                            "TIF",
                            "TEKKOM",
                            "NULL"
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe Updated"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe.updated@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "mahasiswa",
                            "admin",
                            "kaprodi",
                            "dosen"
                        ]
                    },
                    "program_studi": {
                        "type": "string",
                        "enum": [
                            "PTI",
                            "SI",
                            "TI",
                            "TIF",
                            "TEKKOM",
                            "NULL"
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}